requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
streamlit==1.25.0
plotly==5.15.0
joblib==1.3.1
xgboost==1.7.6
kafka-python==2.0.2
pyspark==3.4.1
hdfs==2.7.0
pyarrow==12.0.1